Massive cleanup.
Still WIP but removed a ton of irrelevant stuff and began working on getting to a usable state.

Deep actions are gone for now. they will serve a completely different function later.

No more scriptable objects (for now). Deep behaviors are intended to be entirely defined through code.

This leaves the problem of asset referencing, but I think we should be able to figure something out.