Behaviors kind of done.
In a previous version of this (not on the git) I needed to do a lot of extra stuff when removing behaviors to prevent weird problems from happening. I think because I am using delegates that should not be an issue.

Up next will probably be gutting the cool attribute <-> resource interactions. Attributes will still effect resources, but I think it will be baked in. no more inspector