Entities are updated through a manager
This adds a little more structure to the system and technically performs a little better with huge numbers of entities.

UPDATE is the main logic function that everything should happen on.

LATEUPDATE is when killed entities are removed

Events like "onDeath" still get called during UPDATE.

Updated the readme a bit....